requirements.txt
motor
python-dotenv
fastapi[standard]
pydantic
uvicorn
httpx 
google-genai
tenacity
langchain